Added missing likelihood terms in pgibbs.
At present then a CSMC sweep simply ignores any of the global variables
in terms of the probabilistic semantics.  This is fine when they are
defined unconditionally in the model (hence the tests have been passing)
but incorrect when they are sampled conditioned on other variables.
When this is the case there are extra likelihood terms - we need to
factor the trace probability using the old assume statement.  For
example, if we a model that defines p(x1,x2,y1,y2) =
p(x1)p(y1|x1)p(x2|x1)p(y2|x1,x2) and we do a CSMC sweep fixing x2 then
we are targeting p(x1|x2,y1,y2) \propto p(x1,x2,y1,y2) and so we still
need the old p(x2|x1) term which is now a likelihood.  Note that this
has slipped through the net because if you instead fix x1, then the
p(x1) term is a constant across particles and so can be ignored.  In
general though we cannot as p(x2|x1) is different for each particle
because of having different x1s.

We should probably also add a test to confirm this case. It is
interesting that the gibbs.jl/gibbs.jl test does not catch the bug as it
should manifest for this problem.  My hypothesis for what is happening
here is that even with the bug you might concevieably get the correct
marginal distributions for the variables, but not the right joint
distribution - each gibbs could actually end up targeting the marginal
rather than the conditional.  This needs more investigation though